chore: move traits parsing into submodule
