+++ compiler/noirc_frontend/src/parser/parser.rs
@@ -40,9 +40,8 @@ use crate::parser::{force, ignore_then_commit, statement_recovery};
 use crate::token::{Attribute, Attributes, Keyword, SecondaryAttribute, Token, TokenKind};
 use crate::{
     BinaryOp, BinaryOpKind, BlockExpression, Distinctness, ForLoopStatement, ForRange,
-    FunctionReturnType, FunctionVisibility, Ident, IfExpression, InfixExpression, LValue, Lambda,
-    Literal, NoirStruct, NoirTrait, NoirTraitImpl, NoirTypeAlias, Param, Path, Pattern,
-    Recoverable, Statement, TraitBound, TraitImplItem, TraitItem, TypeImpl,
+    FunctionReturnType, Ident, IfExpression, InfixExpression, LValue, Lambda, Literal, NoirStruct,
+    NoirTypeAlias, Param, Path, Pattern, Recoverable, Statement, TraitBound, TypeImpl,
     UnresolvedTraitConstraint, UnresolvedTypeExpression, UseTree, UseTreeKind, Visibility,
 };
 
@@ -54,6 +53,7 @@ mod assertion;
 mod function;
 mod path;
 mod primitives;
+mod traits;
 
 use path::{maybe_empty_path, path};
 use primitives::{dereference, ident, negation, not, nothing, right_shift_operator, token_kind};
@@ -120,8 +120,8 @@ fn top_level_statement(
     choice((
         function::function_definition(false).map(TopLevelStatement::Function),
         struct_definition(),
-        trait_definition(),
-        trait_implementation(),
+        traits::trait_definition(),
+        traits::trait_implementation(),
         implementation(),
         type_alias_definition().then_ignore(force(just(Token::Semicolon))),
         submodule(module_parser.clone()),
@@ -295,63 +295,6 @@ fn self_parameter() -> impl NoirParser<Param> {
         })
 }
 
-fn trait_definition() -> impl NoirParser<TopLevelStatement> {
-    keyword(Keyword::Trait)
-        .ignore_then(ident())
-        .then(function::generics())
-        .then(where_clause())
-        .then_ignore(just(Token::LeftBrace))
-        .then(trait_body())
-        .then_ignore(just(Token::RightBrace))
-        .map_with_span(|(((name, generics), where_clause), items), span| {
-            TopLevelStatement::Trait(NoirTrait { name, generics, where_clause, span, items })
-        })
-}
-
-fn trait_body() -> impl NoirParser<Vec<TraitItem>> {
-    trait_function_declaration()
-        .or(trait_type_declaration())
-        .or(trait_constant_declaration())
-        .repeated()
-}
-
-fn optional_default_value() -> impl NoirParser<Option<Expression>> {
-    ignore_then_commit(just(Token::Assign), expression()).or_not()
-}
-
-fn trait_constant_declaration() -> impl NoirParser<TraitItem> {
-    keyword(Keyword::Let)
-        .ignore_then(ident())
-        .then_ignore(just(Token::Colon))
-        .then(parse_type())
-        .then(optional_default_value())
-        .then_ignore(just(Token::Semicolon))
-        .validate(|((name, typ), default_value), span, emit| {
-            emit(ParserError::with_reason(
-                ParserErrorReason::ExperimentalFeature("Associated constants"),
-                span,
-            ));
-            TraitItem::Constant { name, typ, default_value }
-        })
-}
-
-/// trait_function_declaration: 'fn' ident generics '(' declaration_parameters ')' function_return_type
-fn trait_function_declaration() -> impl NoirParser<TraitItem> {
-    let trait_function_body_or_semicolon =
-        block(fresh_statement()).map(Option::from).or(just(Token::Semicolon).to(Option::None));
-
-    keyword(Keyword::Fn)
-        .ignore_then(ident())
-        .then(function::generics())
-        .then(parenthesized(function_declaration_parameters()))
-        .then(function_return_type().map(|(_, typ)| typ))
-        .then(where_clause())
-        .then(trait_function_body_or_semicolon)
-        .map(|(((((name, generics), parameters), return_type), where_clause), body)| {
-            TraitItem::Function { name, generics, parameters, return_type, where_clause, body }
-        })
-}
-
 fn validate_attributes(
     attributes: Vec<Attribute>,
     span: Span,
@@ -430,19 +373,6 @@ fn function_declaration_parameters() -> impl NoirParser<Vec<(Ident, UnresolvedTy
         .labelled(ParsingRuleLabel::Parameter)
 }
 
-/// trait_type_declaration: 'type' ident generics
-fn trait_type_declaration() -> impl NoirParser<TraitItem> {
-    keyword(Keyword::Type).ignore_then(ident()).then_ignore(just(Token::Semicolon)).validate(
-        |name, span, emit| {
-            emit(ParserError::with_reason(
-                ParserErrorReason::ExperimentalFeature("Associated types"),
-                span,
-            ));
-            TraitItem::Type { name }
-        },
-    )
-}
-
 /// Parses a non-trait implementation, adding a set of methods to a type.
 ///
 /// implementation: 'impl' generics type '{' function_definition ... '}'
@@ -456,61 +386,6 @@ fn implementation() -> impl NoirParser<TopLevelStatement> {
         .map(|((generics, (object_type, type_span)), methods)| {
             TopLevelStatement::Impl(TypeImpl { generics, object_type, type_span, methods })
         })
-}
-
-/// Parses a trait implementation, implementing a particular trait for a type.
-/// This has a similar syntax to `implementation`, but the `for type` clause is required,
-/// and an optional `where` clause is also useable.
-///
-/// trait_implementation: 'impl' generics ident generic_args for type '{' trait_implementation_body '}'
-fn trait_implementation() -> impl NoirParser<TopLevelStatement> {
-    keyword(Keyword::Impl)
-        .ignore_then(function::generics())
-        .then(path())
-        .then(generic_type_args(parse_type()))
-        .then_ignore(keyword(Keyword::For))
-        .then(parse_type())
-        .then(where_clause())
-        .then_ignore(just(Token::LeftBrace))
-        .then(trait_implementation_body())
-        .then_ignore(just(Token::RightBrace))
-        .map(|args| {
-            let ((other_args, where_clause), items) = args;
-            let (((impl_generics, trait_name), trait_generics), object_type) = other_args;
-
-            TopLevelStatement::TraitImpl(NoirTraitImpl {
-                impl_generics,
-                trait_name,
-                trait_generics,
-                object_type,
-                items,
-                where_clause,
-            })
-        })
-}
-
-fn trait_implementation_body() -> impl NoirParser<Vec<TraitImplItem>> {
-    let function = function::function_definition(true).validate(|mut f, span, emit| {
-        if f.def().is_internal
-            || f.def().is_unconstrained
-            || f.def().is_open
-            || f.def().visibility != FunctionVisibility::Private
-        {
-            emit(ParserError::with_reason(ParserErrorReason::TraitImplFunctionModifiers, span));
-        }
-        // Trait impl functions are always public
-        f.def_mut().visibility = FunctionVisibility::Public;
-        TraitImplItem::Function(f)
-    });
-
-    let alias = keyword(Keyword::Type)
-        .ignore_then(ident())
-        .then_ignore(just(Token::Assign))
-        .then(parse_type())
-        .then_ignore(just(Token::Semicolon))
-        .map(|(name, alias)| TraitImplItem::Type { name, alias });
-
-    function.or(alias).repeated()
 }
 
 fn where_clause() -> impl NoirParser<Vec<UnresolvedTraitConstraint>> {
@@ -1724,33 +1599,6 @@ mod test {
                 "for i in 0...100 {}", // Only '..' is supported, there are no inclusive ranges yet
                 "for i in 0..=100 {}", // Only '..' is supported, there are no inclusive ranges yet
             ],
-        );
-    }
-
-    #[test]
-    fn parse_trait() {
-        parse_all(
-            trait_definition(),
-            vec![
-                // Empty traits are legal in Rust and sometimes used as a way to whitelist certain types
-                // for a particular operation. Also known as `tag` or `marker` traits:
-                // https://stackoverflow.com/questions/71895489/what-is-the-purpose-of-defining-empty-impl-in-rust
-                "trait Empty {}",
-                "trait TraitWithDefaultBody { fn foo(self) {} }",
-                "trait TraitAcceptingMutableRef { fn foo(&mut self); }",
-                "trait TraitWithTypeBoundOperation { fn identity() -> Self; }",
-                "trait TraitWithAssociatedType { type Element; fn item(self, index: Field) -> Self::Element; }",
-                "trait TraitWithAssociatedConstant { let Size: Field; }",
-                "trait TraitWithAssociatedConstantWithDefaultValue { let Size: Field = 10; }",
-                "trait GenericTrait<T> { fn elem(&mut self, index: Field) -> T; }",
-                "trait GenericTraitWithConstraints<T> where T: SomeTrait { fn elem(self, index: Field) -> T; }",
-                "trait TraitWithMultipleGenericParams<A, B, C> where A: SomeTrait, B: AnotherTrait<C> { let Size: Field; fn zero() -> Self; }",
-            ],
-        );
-
-        parse_all_failing(
-            trait_definition(),
-            vec!["trait MissingBody", "trait WrongDelimiter { fn foo() -> u8, fn bar() -> u8 }"],
         );
     }
